Divide tests/test_util.py into tests/util/....py.
