+++ tests/util/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2023 MosaicML Streaming authors
+# SPDX-License-Identifier: Apache-2.0
