+++ tests/util/test_shorthand.py
